Default to Landscape Orientation (#47)
* Converting GalaxyExplorer from a D3D app to a XAML based app

This updates the project file to a base XAML application instead of a base D3D application.  Future changes to input and menuing will require XAML features on non-HoloLens UAP platforms.

* Converting GalaxyExplorer from a D3D app to a XAML based app

This updates the project file to a base XAML application instead of a base D3D application.  Future changes to input and menuing will require XAML features on non-HoloLens UAP platforms.

* Default to Lanscape Orientation

Default the UnityEngine Screen orientation to landscape.  The mobile platform respects this value and will show GalaxyExplorer landscape on a phone.  It has no effect on some platforms like Windows in desktop mode and HoloLens which is just fine.